fix: move property_scorer in separate file to solve import cycle, fix tests
