+++ .gitignore
@@ -141,3 +141,6 @@ test
 .DS_Store
 sandbox.py
 oracle
+
+# local tests
+_main_*
